Removed warnings from code caused by long-int type mismatches

+++ hw/components/mem/mem_device.cpp
@@ -40,7 +40,7 @@ mem_device::mem_device (const char *_name, unsigned int _size) : slave_device (_
     mem = new unsigned char [size];
     memset (mem, 0, size);
 
-    printf("mem_device: Memory area location: 0x%08x\n", (long)mem);
+    printf("mem_device: Memory area location: 0x%08x\n", (int)mem);
 
 }
